backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
requests
httpx
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
APScheduler
pytz
passlib[bcrypt]
python-jose[cryptography]
scipy
ta
binance-connector
